% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.IDB.R
\name{read.IDB}
\alias{read.IDB}
\title{Read InfoStat and InfoGen files}
\usage{
read.IDB(file)
}
\arguments{
\item{file}{The file to be read, must be created by InfoStat/InfoGen or R.}
}
\value{
A data frame.
}
\description{
Read InfoStat IDB2 and InfoGen IGDB files.
}
