% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.IDB.R
\name{write.IDB}
\alias{write.IDB}
\title{Write InfoStat and InfoGen files}
\usage{
write.IDB(x, file, sep = ".", dec = 2)
}
\arguments{
\item{x}{The object to be written, must be a data frame.}

\item{file}{The file to create. Remember that InfoStat handles .idb2 and InfoGen .igdb files.}

\item{sep}{The separators to use in numerical (real) vectors. Default is "."}

\item{dec}{The number of decimal positions to write. Default is 2.}
}
\description{
Write InfoStat IDB2 and InfoGen IGDB files.
}
